Make VS Code show color picker for tag colors in settings.json editor.
In JSON schema declaration of setting "better-comments.tags" in
package.json, declare the "color" property as VS Code color setting.

Explanation:

In JSON schema docs, I quickly found info on how to declare properties
of objects that are elements of an array:
Add to the object describing the array property:
"items": {
  "type": "object",
  "properties": {
    "<name of property>": <object describing property>,
    ...
  }
}

"color" is "type": "string"; JSON schema doesn't have a concept of
colors, this is a VS Code extension (VS Code uses a superset of JSON
schema according to the docs).
However, the VS Code docs don't explain how to declare a setting as
color (or at least I didn't find it). It seems that you are supposed to
define colors in color schemes only and merely override them in
settings.json.

I searched the web and found
https://stackoverflow.com/questions/68594234/how-do-you-display-the-vscode-color-picker-in-json-regardless-of-the-key
which lead me to
https://github.com/wraith13/vscode-schemas/blob/master/en/latest/schemas/token-styling.json.
There, I found the crucial setting "format": "color-hex".